Vision & other useful things (#11)
* Update build.gradle to use the new gradlerio plugin

* Move java plugin above gradlerio plugin in an attempt to fix missing build task error

* Move plugins into plugin block

* Modify build.gradle to try and fix ci error

I can't run gradle right now (at school), so I'm commiting and testing on the CI server. Commented out deploy stuff to try and replicate the gradlerio sample build.gradle in order to see if that's the problem.

* Modify build.gradle to try and fix ci error

I can't run gradle right now (at school), so I'm commiting and testing on the CI server. Commented out deploy stuff to try and replicate the gradlerio sample build.gradle in order to see if that's the problem.

* Add build task

This change is kind of dumb, but travis is complaining about the lack of a build task...?

* Revert "Add build task"

This reverts commit 0a5ae11239a682521379289a7ba44b12d8d61099.

* Move java plugin above GradleRIO plugin

Since the java plugin defines the build task, moving it up should define it for the GradleRIO plugin...?

* Uncomment everything in build.gradle because it's unrelated to the issue

* Working GradleRIO plugin!

* Kind of working

Added some test code. gradlew build works, but deploy still needs some work.

* Fix CI issue with publishing block

The jar block was using the publishing task, then the publishing block was modifying it... *facepalm*.

* Kind of get navX to work

Some methods still cause crashes, but that doesn't appear to be GradleRIO's fault.

* Do something to fix a segmentation fault

For some reason this fixes a crash that happens with more than half of the sample navX data monitor example code.

This makes no sense.

* Update GradleRIO plugin

* Remove crappy, temporary, 'fake' fix to ntcore crash

Looks like 2017.0.7 fixed it!

* Add some mecanum stuff

* Modified DriveMotorCommand to attempt to diagnose a possible motor assignment problem

* fixed silly bug

* removed duplicate code

* Field oriented mecanum

* Merge field and robot oriented mecanum

* attempt at vision processing onboard

* another attempt at vision processing onboard

* Added commands with todo comments to implement certain autonomous behaviors. Removed examples. Refactored Vision subsystem.

* refactor rename. And fix small bug.

* refactor rename.

* added stub for climbing subsystem

* added stub for fuel subsystem

* moved talon assignments into RobotMap constants

* added DriveBaseBreakInCommand

* A variety of vision things

Note that this removes vision processing markup as an accessible stream. Some cleaning up of this could  be done, like the old commands that are no longer needed.